Let the player use the repair table directly.
Right-clicking when something's already in the repair table will take it
out. Otherwise, right-clicking with a tool in your hand will put it in.

It should be difficult, if not impossible, to leave any amount of liquid
already in the table -- the table should only accept liquid when a tool
is inside, and should stop accepting and use any leftover when a tool is
taken out. However:

RedPower2 fluid pipes in 1.4.7 were known to be particularly mean and
bypass the repair table (i.e. TankContainer) and talk directly to the
inner tanks. This is a really bad thing to do, since the repair table is
what decides what is allowed in the tank. When RP2 eventually updates to
whatever Minecraft is current, it will need to be carefully tested, just
in case it does something silly like that again.